extract devtools-proto crate
