+++ Cargo.toml
@@ -27,6 +27,8 @@ ureq = { version = "0.11", optional = true }
 directories = { version = "2.0", optional = true }
 zip = { version = "0.5", optional = true }
 
+devtools_proto = { version = "0.3", path = "devtools-proto" }
+
 [target.'cfg(windows)'.dependencies]
 winreg = "0.6"
 
@@ -44,3 +46,6 @@ path = "src/lib.rs"
 default = [ "fetch" ]
 fetch = [ "ureq", "directories", "zip" ]
 nightly = []
+
+[workspace]
+members = ["devtools-proto"]
