+++ devtools-proto/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "devtools_proto"
+version = "0.3.0"
+authors = ["Alistair Roche <alistair@sunburnt.country>"]
+edition = "2018"
+description = "Control Chrome programatically"
+license = "MIT"
+homepage = "https://github.com/atroche/rust-headless-chrome"
+repository = "https://github.com/atroche/rust-headless-chrome"
+readme = "README.md"
+
+[dependencies]
+failure = "0.1"
+serde = { version = "1", features = ["serde_derive"] }
+serde_json = "1"
+serde_derive = "1"
