+++ devtools-proto/src/profiler.rs
@@ -43,7 +43,7 @@ pub struct ScriptCoverage {
 pub mod methods {
     use serde::{Deserialize, Serialize};
 
-    use crate::protocol::Method;
+    use crate::Method;
 
     #[derive(Serialize, Debug)]
     #[serde(rename_all = "camelCase")]
