+++ devtools-proto/src/runtime.rs
@@ -1,7 +1,7 @@
 pub mod methods {
     use serde::{Deserialize, Serialize};
 
-    use crate::protocol::Method;
+    use crate::Method;
 
     #[derive(Deserialize, Debug, Clone)]
     #[serde(rename_all = "camelCase")]
