+++ devtools-proto/src/target.rs
@@ -100,7 +100,7 @@ pub mod events {
 pub mod methods {
     use serde::{Deserialize, Serialize};
 
-    use crate::protocol::Method;
+    use crate::Method;
 
     #[derive(Serialize, Debug)]
     pub struct GetTargets {}
